update cors config to allow all origins
