+++ Backend/pi2/src/main/java/com/example/pi2/feign/UserClient.java
@@ -1,11 +1,11 @@
 package com.example.pi2.feign;
 
 import com.example.pi2.domain.UserDto;
-import com.example.pi2.security.config.FeignAuthInterceptor;
+import com.example.pi2.config.security.FeignAuthInterceptor;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.*;
 
-@FeignClient(name = "ms-users", url = "http://localhost:8090/api/v1", configuration = FeignAuthInterceptor.class)
+@FeignClient(name = "ms-users", url = "http://ms-gateway/api/v1", configuration = FeignAuthInterceptor.class)
 public interface UserClient {
 
     @GetMapping("/users/search")
