--- Backend/pi2/src/main/java/com/example/pi2/security/config/FeignAuthInterceptor.java
@@ -1,20 +1,0 @@
-package com.example.pi2.security.config;
-
-import feign.RequestInterceptor;
-import feign.RequestTemplate;
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-@Component
-public class FeignAuthInterceptor implements RequestInterceptor {
-
-    @Override
-    public void apply(RequestTemplate requestTemplate) {
-        requestTemplate.header("Authorization", getBearerTokenHeader());
-    }
-
-    private static String getBearerTokenHeader() {
-        return ((ServletRequestAttributes) RequestContextHolder.getRequestAttributes()).getRequest().getHeader("Authorization");
-    }
-}
